requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
alembic
spacy>=3.7.0